+++ src/components/categoriesMain.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Categories = () => (
+  <div className="w-100">
+    <h1 className="text-align-self">Site Under Construction</h1>
+  </div>
+);
+
+export default Categories;
